+++ honeycomb/meta.py
@@ -32,7 +32,7 @@ storage_type_specs = {
                 "'org.apache.hadoop.hive.serde2.JsonSerDe'\n"
                 "STORED AS TEXTFILE")
     },
-    'pq': {
+    'parquet': {
         # Hive expects timestamps to be saved as 96-bit integers with Parquet,
         # even though this behavior is no longer the Parquet standard.
         'settings': {
@@ -41,6 +41,10 @@ storage_type_specs = {
             'use_deprecated_int96_timestamps': True
         },
         'ddl': 'STORED AS PARQUET'
+    },
+    'orc': {
+        'settings': {},
+        'ddl': 'STORED AS ORC'
     }
 }
 
@@ -111,24 +115,33 @@ def generate_s3_filename(storage_type=None):
     """
     filename = datetime.strftime(datetime.now(), '%Y-%m-%d_%H-%M-%S')
     if storage_type is None:
-        storage_type = 'pq'
+        storage_type = 'parquet'
 
     return '.'.join([filename, storage_type])
 
 
-def get_table_column_order(table_name, schema):
+def get_table_column_order(table_name, schema, include_dtypes=False):
     """
-    Gets the order of columns in a data lake table
+    Gets the order of columns in a data lake table. Deliberately leaves
+    out partition columns, because those should not be present in the original
+    dataset.
 
     Args:
         table_name (str): The table to get the column order of
         schema (str): The schema the table is in
     """
     colname_col = 'col_name'
+    dtype_col = 'data_type'
     description = describe_table(table_name, schema, include_metadata=True)
     colname_end = description.index[description[colname_col] == ''][0] - 1
-    columns = description.loc[:colname_end, colname_col]
-    return columns.to_list()
+
+    description = description.loc[:colname_end]
+    if not include_dtypes:
+        columns = description[colname_col]
+        return columns.to_list()
+    else:
+        columns_w_types = description[[colname_col, dtype_col]]
+        return columns_w_types.rename(columns={dtype_col: 'dtype'})
 
 
 def get_table_metadata(table_name, schema):
@@ -194,7 +207,8 @@ def get_table_storage_type(table_name, schema):
     hive_input_format_to_storage_type = {
         'org.apache.hadoop.hive.ql.io.avro.AvroContainerInputFormat': 'avro',
         'org.apache.hadoop.mapred.TextInputFormat': 'text',
-        'org.apache.hadoop.hive.ql.io.parquet.MapredParquetInputFormat': 'pq',
+        'org.apache.hadoop.hive.ql.io.parquet.MapredParquetInputFormat':
+            'parquet',
         'org.apache.hadoop.hive.ql.io.orc.OrcInputFormat': 'orc'
     }
     format_label_idx = table_metadata.index[