+++ test/test_create_table.py
@@ -2,7 +2,7 @@ import pytest
 
 import river as rv
 
-from honeycomb.create_table import create_table_from_df
+from honeycomb.create_table.create_table_from_df import create_table_from_df
 
 
 def test_create_table_from_df_csv(mocker, setup_bucket_wo_contents,
@@ -12,8 +12,9 @@ def test_create_table_from_df_csv(mocker, setup_bucket_wo_contents,
     for presence of the DataFrame at the expected location in S3
     """
     schema = 'experimental'
-    mocker.patch.dict('honeycomb.create_table.schema_to_zone_bucket_map',
-                      {schema: test_bucket}, clear=True)
+    mocker.patch.dict(
+        'honeycomb.create_table.common.schema_to_zone_bucket_map',
+        {schema: test_bucket}, clear=True)
 
     mocker.patch('honeycomb.hive.run_lake_query', return_value=False)
     mocker.patch('honeycomb.check.table_existence', return_value=False)